refactored ServerParse classes
Refactored the classes to eliminate duplicates and make the code easier to read
